Add dtype to keras_core.operations.one_hot (#327)
* Add dtype to keras_core.operations.one_hot

* Default dtype to None and use `backend.floatx()` instead

* Use `backend.floatx()` in the __init__ itself

+++ keras_core/backend/jax/nn.py
@@ -387,8 +387,8 @@ def conv_transpose(
     )
 
 
-def one_hot(x, num_classes, axis=-1):
-    return jnn.one_hot(x, num_classes, axis=axis)
+def one_hot(x, num_classes, axis=-1, dtype="float32"):
+    return jnn.one_hot(x, num_classes, axis=axis, dtype=dtype)
 
 
 def categorical_crossentropy(target, output, from_logits=False, axis=-1):
